Split HTML lib into several helper files
